Remove stored pwn*.xz files; use stored metadata
The omw-en and omw-en31 lexicons are now generated from WordNet
sources directly, so the stored xml.xz files are no longer
needed. This commit also repurposes, and renames, those files'
directories for storing additional files related to the wordnets, such
as the LICENSE, README.md, and citation.bib files. This commit also
updates and expands the READMEs a bit.