Hard-code syntactic frames
Earlier versions of the Princeton WordNet did not include
verb.Framestext, and they never changed across versions, so it's
easier to just hard-code them than to load them from a file. The only
potential issue I see is that this is content copied from the
copyrighted WordNet documentation and there might not be enough
attribution. I do link back to the documentation, so hopefully we're
good there.