Add wndb.py and build_senseidx.py
build_senseidx.py will create exact replicas of the index.sense files
for WordNet 1.7 and higher versions.

For WordNet 1.6, you can get close with the --use-adjposition option,
but the counts for any sense key with an adjposition (a) or (p) after
the head word of satellite adjective sense keys would need to be reset
to 0.

WordNet 1.5 did not have an index.sense file distributed with it.