Remove blank lemmas/defs/exs with clean.sh
A previous commit removed these lines, but this commit changes the
scripts/clean-tsv.py script to do the same thing and generates records
for the *-changes.tab files. I verified that this reproduced the
hand-committed changes by copying in earlier versions of the affected
files and running the cleaning script. The result had no diffs with
what's in Git and I was able to get the change records.

This commit also changes the clean.sh script so it appends to
the *-changes.tab files instead of overwriting them.