Handle lexical gaps marked by GAP! or PSEUDOGAP!
If a synset has a member GAP! or PSEUDOGAP!, that sense is removed and all
remaining senses are checked with heuristics to see if they are compositional
collocations (the presence of a space or apostrophe for Italian or a space for
Hebrew). If so, the sense is marked `lexicalized=False`. If all senses are
`lexicalized=False`, or there are no other members, the synset is marked
`lexicalized=False`.

Resolves #45